backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
google-generativeai
pydantic-settings
pdf2image